+++ tests/server/blueprints/alignviewers/test_alignviewers_controllers.py
@@ -2,11 +2,65 @@
 from flask_login import current_user
 from flask import url_for
 from scout.server.extensions import cloud_tracks, store
+from scout.commands.update.individual import individual as ind_cmd
 from scout.server.blueprints.alignviewers import controllers
 
 
+def test_make_sashimi_tracks(app, case_obj):
+    """Test function that creates splice junction tracks"""
+    # GIVEN a case's affected individual
+    affected_name = "NA12882"
+
+    # GIVEN a case with an individual containing splice junction data:
+    runner = app.test_cli_runner()
+    runner.invoke(
+        ind_cmd,
+        [
+            "--case-id",
+            case_obj["_id"],
+            "--ind",
+            affected_name,
+            "rna_coverage_bigwig",
+            "test.bigWig",
+        ],
+        input="y",
+    )
+    runner.invoke(
+        ind_cmd,
+        ["--case-id", case_obj["_id"], "--ind", affected_name, "splice_junctions_bed", "test.bed"],
+        input="y",
+    )
+    updated_case = store.case_collection.find_one()
+    for ind in updated_case["individuals"]:
+        if ind["display_name"] == affected_name:
+            assert ind["rna_coverage_bigwig"]
+            assert ind["splice_junctions_bed"]
+
+    with app.test_client() as client:
+        # GIVEN that the user could be logged in
+        resp = client.get(url_for("auto_login"))
+        assert resp.status_code == 200
+
+        # WHEN the gene splice junction track is created for any variant in a gene
+        test_variant = store.variant_collection.find_one({"hgnc_symbols": ["POT1"]})
+
+        # THEN it should return the expected data
+        display_obj = controllers.make_sashimi_tracks(
+            case_obj["owner"], case_obj["display_name"], test_variant["_id"], "37"
+        )
+        assert display_obj["case"] == case_obj["display_name"]
+        assert display_obj["genes"] == ["POT1"]
+        assert display_obj["locus"]
+        assert display_obj["tracks"][0]["name"] == affected_name
+        assert display_obj["tracks"][0]["coverage_wig"] == "test.bigWig"
+        assert display_obj["tracks"][0]["splicej_bed"] == "test.bed"
+        assert display_obj["tracks"][0]["splicej_bed_index"] is None
+        assert display_obj["reference_track"]  # Reference genome track
+        assert display_obj["custom_tracks"]  # Custom tracks include gene track in the right build
+
+
 def test_set_cloud_public_tracks(app):
-    """ Test function that adds cloud public tracks to track display object"""
+    """Test function that adds cloud public tracks to track display object"""
 
     # GIVEN an app with public cloud tracks initialized
     patched_track = {"37": [{"name": "test track"}]}