+++ CHANGELOG.md
@@ -10,6 +10,8 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Show variant allele frequencies for mitochondrial variants
 - Extend "public" json API with diseases (OMIM) and phenotypes (HPO)
 - HPO gene list download now has option for clinical and non-clinical genes
+- Display gene splice junctions data in sashimi plots
+- Update case individuals with splice junctions tracks
 ### Fixed
 - Show other causative once, even if several events point to it
 - Filtering variants by mitochondrial chromosome for cases with genome build=38